# openapi/openapi.yaml
...
openapi: "3.1.0"
info:
  title: "Toolforge jobs framework API"
  version: "2.0.0"
servers:
  - url: "http://127.0.0.1:8000"
    description: "Local direct development server."
  - url: "https://127.0.0.1:3000/jobs"
    description: "Lima-kilo development server."
  - url: "https://api.svc.tools.eqiad1.wikimedia.cloud:30003/jobs"
    description: "Toolforge internal API gateway endpoint."
  - url: "https://api.svc.toolsbeta.eqiad1.wikimedia.cloud:30003/jobs"
    description: "Toolforge beta internal API gateway endpoint."
security:
  - toolforge_tool_client_certificate: []
paths:
...
  /api/v1/jobs:
    get:
      operationId: "list"
      description: "Returns the list of jobs defined for this Toolforge tool account."
      responses:
        "200":
          description: "The list of jobs is returned."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DefinedJob"
        "401":
          $ref: "#/components/responses/unauthorized_error"
    post:
      operationId: "create"
      description: "Create a new job in this Toolforge tool account."
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/NewJob"
      responses:
        "200":
          description: "A new job was created."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefinedJob"
        "400":
          $ref: "#/components/responses/bad_parameters_error"
        "401":
          $ref: "#/components/responses/unauthorized_error"
    delete:
      operationId: "flush"
      description: "Stop and delete all defined jobs from this Toolforge tool account."
      responses:
        "200":
          description: "All jobs were flushed"
        "401":
          $ref: "#/components/responses/unauthorized_error"
  /api/v1/tool/{toolname}/jobs:
    get:
      operationId: "list_with_toolname"
      description: "Returns the list of jobs defined for this Toolforge tool account."
      parameters:
        - name: toolname
          in: path
          description: "Tool name."
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "The list of jobs is returned."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DefinedJob"
        "401":
          $ref: "#/components/responses/unauthorized_error"
    post:
      operationId: "create_with_toolname"
      description: "Create a new job in this Toolforge tool account."
      parameters:
        - name: toolname
          in: path
          description: "Tool name."
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/NewJob"
      responses:
        "200":
          description: "A new job was created."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefinedJob"
        "400":
          $ref: "#/components/responses/bad_parameters_error"
        "401":
          $ref: "#/components/responses/unauthorized_error"
    delete:
      operationId: "flush_with_toolname"
      description: "Stop and delete all defined jobs from this Toolforge tool account."
      parameters:
        - name: toolname
          in: path
          description: "Tool name."
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "All jobs were flushed"
        "401":
          $ref: "#/components/responses/unauthorized_error"
  /api/v1/jobs/{id}:
    get:
      operationId: "show"
      description: "Show information about a given job for this Toolforge tool account."
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: "Job was found."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefinedJob"
        "400":
          $ref: "#/components/responses/bad_parameters_error"
        "401":
          $ref: "#/components/responses/unauthorized_error"
        "404":
          $ref: "#/components/responses/not_found_error"
    delete:
      operationId: "delete"
      description: "Stop and delete the specified job from this Toolforge tool account."
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: "Job was deleted."
        "400":
          $ref: "#/components/responses/bad_parameters_error"
        "401":
          $ref: "#/components/responses/unauthorized_error"
        "404":
          $ref: "#/components/responses/not_found_error"
  /api/v1/tool/{toolname}/jobs/{id}:
    get:
      operationId: "show_with_toolname"
      description: "Show information about a given job for this Toolforge tool account."
      parameters:
        - name: toolname
          in: path
          description: "Tool name."
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: "Job was found."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefinedJob"
        "400":
          $ref: "#/components/responses/bad_parameters_error"
        "401":
          $ref: "#/components/responses/unauthorized_error"
        "404":
          $ref: "#/components/responses/not_found_error"
    delete:
      operationId: "delete_with_toolname"
      description: "Stop and delete the specified job from this Toolforge tool account."
      parameters:
        - name: toolname
          in: path
          description: "Tool name."
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: "Job was deleted."
        "400":
          $ref: "#/components/responses/bad_parameters_error"
        "401":
          $ref: "#/components/responses/unauthorized_error"
        "404":
          $ref: "#/components/responses/not_found_error"
  /api/v1/jobs/{id}/logs:
    get:
      operationId: "get_logs"
      description: "Get logs for a given job in this Toolforge tool account."
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: "Job logs are served."
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/JobLogs"
        "400":
          $ref: "#/components/responses/bad_parameters_error"
        "401":
          $ref: "#/components/responses/unauthorized_error"
        "404":
          $ref: "#/components/responses/not_found_error"
  /api/v1/tool/{toolname}/jobs/{id}/logs:
    get:
      operationId: "get_logs_with_toolname"
      description: "Get logs for a given job in this Toolforge tool account."
      parameters:
        - name: toolname
          in: path
          description: "Tool name."
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: "Job logs are served."
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/JobLogs"
        "400":
          $ref: "#/components/responses/bad_parameters_error"
        "401":
          $ref: "#/components/responses/unauthorized_error"
        "404":
          $ref: "#/components/responses/not_found_error"
  /api/v1/jobs/{id}/restart:
    post:
      operationId: "restart"
      description: "Restart a given job. If the job is a cronjob, execute it right now."
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: "Job has been restarted."
        "400":
          $ref: "#/components/responses/bad_parameters_error"
        "401":
          $ref: "#/components/responses/unauthorized_error"
        "404":
          $ref: "#/components/responses/not_found_error"
  /api/v1/tool/{toolname}/jobs/{id}/restart:
    post:
      operationId: "restart_with_toolname"
      description: "Restart a given job. If the job is a cronjob, execute it right now."
      parameters:
        - name: toolname
          in: path
          description: "Tool name."
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: "Job has been restarted."
        "400":
          $ref: "#/components/responses/bad_parameters_error"
        "401":
          $ref: "#/components/responses/unauthorized_error"
        "404":
          $ref: "#/components/responses/not_found_error"
...
components:
  schemas:
    ResponseMessages:
      type: object
      description: |
        Lists of messages by priority, note that there will be a warning here when the endpoint is being deprecated
        with information on how to proceed to update it.
      properties:
        info:
          type: array
          items:
            type: string
        warning:
          type: array
          items:
            type: string
        error:
          type: array
          items:
            type: string
...
    JobLogs:
      type: string
    ScriptHealthCheck:
      type: object
      description: "Health check implemented as a script that runs inside the container"
      required:
        - type
        - script
      properties:
        type:
          type: string
        script:
          type: string
    # common job fields
    CommonJob:
      type: object
      required:
        - name
        - cmd
      properties:
        name:
          type: string
          description: "Unique name that identifies the job."
          # This is a restriction by Kubernetes:
          # a lowercase RFC 1123 subdomain must consist of lower case alphanumeric
          # characters, '-' or '.', and must start and end with an alphanumeric character
          pattern: "^[a-z0-9]([-a-z0-9]*[a-z0-9])?([.][a-z0-9]([-a-z0-9]*[a-z0-9])?)*$"
          minLength: 1
          maxLength: 52
          example: "my-job"
        cmd:
          type: string
          description: "Command that this job is executing."
          example: "./some-command.sh --with-args"
        filelog:
          type: boolean
          description: "Whether this job uses filelog or not."
          example: false
        filelog_stdout:
          type: string
          description: "Path to the stdout file log."
          example: "logs/my-job.out"
        filelog_stderr:
          type: string
          description: "Path to the stderr file log."
          example: "logs/my-job.err"
        emails:
          type: string
          description: "Job emails setting."
          enum:
            - "none"
            - "all"
            - "onfinish"
            - "onfailure"
          example: "all"
        retry:
          type: integer
          description: "Job retry policy. Zero means don't retry at all (the default)"
          minimum: 0
          maximum: 5
          example: 0
        mount:
          type: string
          description: "NFS mount configuration for the job."
          enum:
            - "all"
            - "none"
          example: "none"
        schedule:
          type: string
          description: "If the job is a cronjob, execution schedule."
          example: "@hourly"
        continuous:
          type: boolean
          description: "If a job should be always running."
          example: false
        port:
          type: integer
          description: "Port to expose for the job."
          minimum: 1
          maximum: 65535
          example: 8080
        cpu:
          type: string
          description: "Job CPU resource limit."
          example: "1"
        health_check:
          type: object
          description: "Health check configuration for the job."
          oneOf:
            - $ref: "#/components/schemas/ScriptHealthCheck"
          discriminator:
            propertyName: type
            mapping:
              script: "#/components/schemas/ScriptHealthCheck"
        memory:
          type: string
          description: "Job memory resource limit."
          example: "1G"
    # job as created by the user
    NewJob:
      allOf:
        - $ref: "#/components/schemas/CommonJob"
        - type: object
          required:
            - imagename
          properties:
            imagename:
              type: string
              description: "Container image the job uses."
              example: "tool-my-tool/tool-my-tool:latest"
    # job, once defined, as reported by the API
    DefinedJob:
      allOf:
        - $ref: "#/components/schemas/CommonJob"
        - type: object
          required:
            - image
            - image_state
            - status_short
            - status_long
          properties:
            image:
              type: string
              description: "Container image the job uses."
              example: "tool-my-tool/tool-my-tool:latest"
            image_state:
              type: string
              description: "Information about the image state, i.e, if it is stable, deprecated, or something else."
              example: "stable"
            status_short:
              type: string
              description: "Job status, short text."
              example: "Running."
            status_long:
              type: string
              description: "Job status, extended text."
              example: "Running since 30 seconds ago."
            schedule_actual:
              type: string
              description: "If the job is a cronjob, actual execution schedule."
              example: "15 * * * *"
  parameters:
    JobId:
      name: id
      in: path
      description: "Job name."
      required: true
      style: simple
      explode: false
      schema:
        type: string
  responses:
    unauthorized_error:
      description: "Unauthorized error."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ResponseMessages"
    bad_parameters_error:
      description: "Bad parameters error"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ResponseMessages"
    not_found_error:
      description: "Object not found error."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ResponseMessages"
  securitySchemes:
    toolforge_tool_client_certificate:
      type: "mutualTLS"
      description: "This API requires a client certificate generated for the Toolforge tool account."